@@ -5,8 +5,8 @@ version: 4.1.4
 homepage: https://github.com/aryzhov/flutter-expandable
 
 environment:
-  sdk: ">=2.7.0 < 3.0.0"
-  flutter: "^1.12.0"
+  sdk: '>=2.12.0 <3.0.0'
+  # flutter: "^1.12.0"
 
 dependencies:
   flutter:
