@@ -32,16 +32,16 @@ class ExpandableThemeData {
   static final ExpandableThemeData empty = ExpandableThemeData();
 
   // Expand icon color.
-  final Color iconColor;
+  final Color? iconColor;
 
   // If true then [InkWell] will be used in the header for a ripple effect.
-  final bool useInkWell;
+  final bool? useInkWell;
 
   // The duration of the transition between collapsed and expanded states.
-  final Duration animationDuration;
+  final Duration? animationDuration;
 
   // The duration of the scroll animation to make the expanded widget visible.
-  final Duration scrollAnimationDuration;
+  final Duration? scrollAnimationDuration;
 
   /// The point in the cross-fade animation timeline (from 0 to 1)
   /// where the [collapsed] and [expanded] widgets are half-visible.
@@ -59,57 +59,57 @@ class ExpandableThemeData {
   ///
   /// When collapsing, the effect of this setting is reversed. For example, if the value is 0
   /// then the [expanded] widget will remain to be shown until the end of the size animation.
-  final double crossFadePoint;
+  final double? crossFadePoint;
 
   /// The alignment of widgets during animation between expanded and collapsed states.
-  final AlignmentGeometry alignment;
+  final AlignmentGeometry? alignment;
 
   // Fade animation curve between expanded and collapsed states.
-  final Curve fadeCurve;
+  final Curve? fadeCurve;
 
   // Size animation curve between expanded and collapsed states.
-  final Curve sizeCurve;
+  final Curve? sizeCurve;
 
   // The alignment of the header for `ExpandablePanel`.
-  final ExpandablePanelHeaderAlignment headerAlignment;
+  final ExpandablePanelHeaderAlignment? headerAlignment;
 
   // The alignment of the body for `ExpandablePanel`.
-  final ExpandablePanelBodyAlignment bodyAlignment;
+  final ExpandablePanelBodyAlignment? bodyAlignment;
 
   /// Expand icon placement.
-  final ExpandablePanelIconPlacement iconPlacement;
+  final ExpandablePanelIconPlacement? iconPlacement;
 
   /// If true, the header of [ExpandablePanel] can be clicked by the user to expand or collapse.
-  final bool tapHeaderToExpand;
+  final bool? tapHeaderToExpand;
 
   /// If true, the body of [ExpandablePanel] can be clicked by the user to expand.
-  final bool tapBodyToExpand;
+  final bool? tapBodyToExpand;
 
   /// If true, the body of [ExpandablePanel] can be clicked by the user to collapse.
-  final bool tapBodyToCollapse;
+  final bool? tapBodyToCollapse;
 
   /// If true, an icon is shown in the header of [ExpandablePanel].
-  final bool hasIcon;
+  final bool? hasIcon;
 
   /// Expand icon size.
-  final double iconSize;
+  final double? iconSize;
 
   /// Expand icon padding.
-  final EdgeInsets iconPadding;
+  final EdgeInsets? iconPadding;
 
   /// Icon rotation angle in clockwise radians. For example, specify `math.pi` to rotate the icon by 180 degrees
   /// clockwise when clicking on the expand button.
-  final double iconRotationAngle;
+  final double? iconRotationAngle;
 
   /// The icon in the collapsed state.
-  final IconData expandIcon;
+  final IconData? expandIcon;
 
   /// The icon in the expanded state. If you specify the same icon as `expandIcon`, the `expandIcon` icon will
   /// be shown upside-down in the expanded state.
-  final IconData collapseIcon;
+  final IconData? collapseIcon;
 
   ///The [BorderRadius] for the [InkWell], if `inkWell` is set to true
-  final BorderRadius inkWellBorderRadius;
+  final BorderRadius? inkWellBorderRadius;
 
   const ExpandableThemeData({
     this.iconColor,
@@ -136,18 +136,19 @@ class ExpandableThemeData {
   });
 
   static ExpandableThemeData combine(
-      ExpandableThemeData theme, ExpandableThemeData defaults) {
+      ExpandableThemeData? theme, ExpandableThemeData? defaults) {
     if (defaults == null || defaults.isEmpty()) {
       return theme ?? empty;
     } else if (theme == null || theme.isEmpty()) {
-      return defaults ?? empty;
+      return defaults;
     } else if (theme.isFull()) {
       return theme;
     } else {
       return ExpandableThemeData(
         iconColor: theme.iconColor ?? defaults.iconColor,
         useInkWell: theme.useInkWell ?? defaults.useInkWell,
-        inkWellBorderRadius: theme.inkWellBorderRadius ?? defaults.inkWellBorderRadius,
+        inkWellBorderRadius:
+            theme.inkWellBorderRadius ?? defaults.inkWellBorderRadius,
         animationDuration:
             theme.animationDuration ?? defaults.animationDuration,
         scrollAnimationDuration:
@@ -177,19 +178,20 @@ class ExpandableThemeData {
 
   // ignore: deprecated_member_use_from_same_package
   double get collapsedFadeStart =>
-      crossFadePoint < 0.5 ? 0 : (crossFadePoint * 2 - 1);
+      crossFadePoint! < 0.5 ? 0 : (crossFadePoint! * 2 - 1);
 
   // ignore: deprecated_member_use_from_same_package
-  double get collapsedFadeEnd => crossFadePoint < 0.5 ? 2 * crossFadePoint : 1;
+  double get collapsedFadeEnd =>
+      crossFadePoint! < 0.5 ? 2 * crossFadePoint! : 1;
 
   // ignore: deprecated_member_use_from_same_package
   double get expandedFadeStart =>
-      crossFadePoint < 0.5 ? 0 : (crossFadePoint * 2 - 1);
+      crossFadePoint! < 0.5 ? 0 : (crossFadePoint! * 2 - 1);
 
   // ignore: deprecated_member_use_from_same_package
-  double get expandedFadeEnd => crossFadePoint < 0.5 ? 2 * crossFadePoint : 1;
+  double get expandedFadeEnd => crossFadePoint! < 0.5 ? 2 * crossFadePoint! : 1;
 
-  ExpandableThemeData nullIfEmpty() {
+  ExpandableThemeData? nullIfEmpty() {
     return isEmpty() ? null : this;
   }
 
@@ -260,7 +262,7 @@ class ExpandableThemeData {
   }
 
   static ExpandableThemeData withDefaults(
-      ExpandableThemeData theme, BuildContext context,
+      ExpandableThemeData? theme, BuildContext context,
       {bool rebuildOnChange = true}) {
     if (theme != null && theme.isFull()) {
       return theme;
@@ -276,11 +278,11 @@ class ExpandableTheme extends StatelessWidget {
   final ExpandableThemeData data;
   final Widget child;
 
-  ExpandableTheme({@required this.data, @required this.child});
+  ExpandableTheme({required this.data, required this.child});
 
   @override
   Widget build(BuildContext context) {
-    _ExpandableThemeNotifier n =
+    _ExpandableThemeNotifier? n =
         context.dependOnInheritedWidgetOfExactType<_ExpandableThemeNotifier>();
     return _ExpandableThemeNotifier(
       themeData: ExpandableThemeData.combine(data, n?.themeData),
@@ -292,21 +294,21 @@ class ExpandableTheme extends StatelessWidget {
 /// Makes an [ExpandableController] available to the widget subtree.
 /// Useful for making multiple [Expandable] widgets synchronized with a single controller.
 class ExpandableNotifier extends StatefulWidget {
-  final ExpandableController controller;
-  final bool initialExpanded;
+  final ExpandableController? controller;
+  final bool? initialExpanded;
   final Widget child;
 
   ExpandableNotifier(
       {
       // An optional key
-      Key key,
+      Key? key,
 
       /// If the controller is not provided, it's created with the initial value of `initialExpanded`.
       this.controller,
 
       /// Initial expanded state. Must not be used together with [controller].
       this.initialExpanded,
-      @required
+      required
 
           /// The child can be any widget which contains [Expandable] widgets in its widget tree.
           this.child})
@@ -318,8 +320,8 @@ class ExpandableNotifier extends StatefulWidget {
 }
 
 class _ExpandableNotifierState extends State<ExpandableNotifier> {
-  ExpandableController controller;
-  ExpandableThemeData theme;
+  ExpandableController? controller;
+  ExpandableThemeData? theme;
 
   @override
   void initState() {
@@ -354,16 +356,16 @@ class _ExpandableNotifierState extends State<ExpandableNotifier> {
 class _ExpandableControllerNotifier
     extends InheritedNotifier<ExpandableController> {
   _ExpandableControllerNotifier(
-      {@required ExpandableController controller, @required Widget child})
+      {required ExpandableController? controller, required Widget child})
       : super(notifier: controller, child: child);
 }
 
 /// Makes an [ExpandableController] available to the widget subtree.
 /// Useful for making multiple [Expandable] widgets synchronized with a single controller.
 class _ExpandableThemeNotifier extends InheritedWidget {
-  final ExpandableThemeData themeData;
+  final ExpandableThemeData? themeData;
 
-  _ExpandableThemeNotifier({@required this.themeData, @required Widget child})
+  _ExpandableThemeNotifier({required this.themeData, required Widget child})
       : super(child: child);
 
   @override
@@ -379,10 +381,10 @@ class ExpandableController extends ValueNotifier<bool> {
   /// Returns [true] if the state is expanded, [false] if collapsed.
   bool get expanded => value;
 
-  final Duration _animationDuration;
+  final Duration? _animationDuration;
 
   ExpandableController(
-      {bool initialExpanded,
+      {bool? initialExpanded,
       @deprecated
           // ignore: deprecated_member_use_from_same_package
           animationDuration})
@@ -402,7 +404,7 @@ class ExpandableController extends ValueNotifier<bool> {
     expanded = !expanded;
   }
 
-  static ExpandableController of(BuildContext context,
+  static ExpandableController? of(BuildContext context,
       {bool rebuildOnChange = true}) {
     final notifier = rebuildOnChange
         ? context
@@ -417,26 +419,26 @@ class ExpandableController extends ValueNotifier<bool> {
 /// The state is determined by an instance of [ExpandableController] provided by [ScopedModel]
 class Expandable extends StatelessWidget {
   /// Whe widget to show when collapsed
-  final Widget collapsed;
+  final Widget? collapsed;
 
   /// The widget to show when expanded
-  final Widget expanded;
+  final Widget? expanded;
 
   /// If the controller is not specified, it will be retrieved from the context
-  final ExpandableController controller;
+  final ExpandableController? controller;
 
-  final ExpandableThemeData _theme;
+  final ExpandableThemeData? _theme;
 
   Expandable({
-    Key key,
+    Key? key,
     this.collapsed,
     this.expanded,
     this.controller,
-    ExpandableThemeData theme,
-    @deprecated double crossFadePoint,
-    @deprecated Curve fadeCurve,
-    @deprecated Curve sizeCurve,
-    @deprecated AlignmentGeometry alignment,
+    ExpandableThemeData? theme,
+    @deprecated double? crossFadePoint,
+    @deprecated Curve? fadeCurve,
+    @deprecated Curve? sizeCurve,
+    @deprecated AlignmentGeometry? alignment,
     // ignore: deprecated_member_use_from_same_package
   })  : _theme = ExpandableThemeData.combine(
                 ExpandableThemeData(
@@ -450,22 +452,22 @@ class Expandable extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final controller = this.controller ?? ExpandableController.of(context);
+    final controller = this.controller ?? ExpandableController.of(context)!;
     final theme = ExpandableThemeData.withDefaults(_theme, context);
 
     return AnimatedCrossFade(
-      alignment: theme.alignment,
+      alignment: theme.alignment!,
       firstChild: collapsed ?? Container(),
       secondChild: expanded ?? Container(),
       firstCurve: Interval(theme.collapsedFadeStart, theme.collapsedFadeEnd,
-          curve: theme.fadeCurve),
+          curve: theme.fadeCurve!),
       secondCurve: Interval(theme.expandedFadeStart, theme.expandedFadeEnd,
-          curve: theme.fadeCurve),
-      sizeCurve: theme.sizeCurve,
+          curve: theme.fadeCurve!),
+      sizeCurve: theme.sizeCurve!,
       crossFadeState: controller.expanded
           ? CrossFadeState.showSecond
           : CrossFadeState.showFirst,
-      duration: controller._animationDuration ?? theme.animationDuration,
+      duration: controller._animationDuration ?? theme.animationDuration!,
     );
   }
 }
@@ -509,39 +511,39 @@ enum ExpandablePanelBodyAlignment {
 /// A configurable widget for showing user-expandable content with an optional expand button.
 class ExpandablePanel extends StatelessWidget {
   /// If specified, the header is always shown, and the expandable part is shown under the header
-  final Widget header;
+  final Widget? header;
 
   /// The widget shown in the collapsed state
-  final Widget collapsed;
+  final Widget? collapsed;
 
   /// The widget shown in the expanded state
-  final Widget expanded;
+  final Widget? expanded;
 
   /// Builds an Expandable object, optional
-  final ExpandableBuilder builder;
+  final ExpandableBuilder? builder;
 
   /// An optional controller. If not specified, a default controller will be
   /// obtained from a surrounding [ExpandableNotifier]. If that does not exist,
   /// the controller will be created with the initial state of [initialExpanded].
-  final ExpandableController controller;
+  final ExpandableController? controller;
 
-  final ExpandableThemeData _theme;
+  final ExpandableThemeData? _theme;
 
   ExpandablePanel({
-    Key key,
+    Key? key,
     this.header,
     this.collapsed,
     this.expanded,
     this.controller,
     this.builder,
     // Theme overrides (optional)
-    @deprecated bool tapHeaderToExpand,
-    @deprecated bool tapBodyToCollapse,
-    @deprecated bool hasIcon,
-    ExpandableThemeData theme,
-    @deprecated ExpandablePanelIconPlacement iconPlacement,
-    @deprecated Color iconColor,
-    @deprecated ExpandablePanelHeaderAlignment headerAlignment,
+    @deprecated bool? tapHeaderToExpand,
+    @deprecated bool? tapBodyToCollapse,
+    @deprecated bool? hasIcon,
+    ExpandableThemeData? theme,
+    @deprecated ExpandablePanelIconPlacement? iconPlacement,
+    @deprecated Color? iconColor,
+    @deprecated ExpandablePanelHeaderAlignment? headerAlignment,
 
     // ignore: deprecated_member_use_from_same_package
   })  : _theme = ExpandableThemeData.combine(
@@ -561,8 +563,8 @@ class ExpandablePanel extends StatelessWidget {
   Widget build(BuildContext context) {
     final theme = ExpandableThemeData.withDefaults(_theme, context);
 
-    Widget buildHeaderRow() {
-      CrossAxisAlignment calculateHeaderCrossAxisAlignment() {
+    Widget? buildHeaderRow() {
+      CrossAxisAlignment? calculateHeaderCrossAxisAlignment() {
         switch (theme.headerAlignment) {
           case ExpandablePanelHeaderAlignment.top:
             return CrossAxisAlignment.start;
@@ -570,43 +572,45 @@ class ExpandablePanel extends StatelessWidget {
             return CrossAxisAlignment.center;
           case ExpandablePanelHeaderAlignment.bottom:
             return CrossAxisAlignment.end;
+          case null:
+            break;
         }
         assert(false);
         return null;
       }
 
-      Widget wrapWithExpandableButton({Widget widget, bool wrap}) {
+      Widget? wrapWithExpandableButton({Widget? widget, required bool wrap}) {
         return wrap ? ExpandableButton(child: widget) : widget;
       }
 
-      if (!theme.hasIcon) {
+      if (!theme.hasIcon!) {
         return wrapWithExpandableButton(
-            widget: header, wrap: theme.tapHeaderToExpand);
+            widget: header, wrap: theme.tapHeaderToExpand!);
       } else {
-        final rowChildren = <Widget>[
+        final rowChildren = <Widget?>[
           Expanded(
-            child: header,
+            child: header!,
           ),
           // ignore: deprecated_member_use_from_same_package
           wrapWithExpandableButton(
               widget: ExpandableIcon(theme: theme),
-              wrap: !theme.tapHeaderToExpand)
+              wrap: !theme.tapHeaderToExpand!)
         ];
         return wrapWithExpandableButton(
             widget: Row(
-              crossAxisAlignment: calculateHeaderCrossAxisAlignment(),
+              crossAxisAlignment: calculateHeaderCrossAxisAlignment()!,
               children:
                   theme.iconPlacement == ExpandablePanelIconPlacement.right
-                      ? rowChildren
-                      : rowChildren.reversed.toList(),
+                      ? rowChildren as List<Widget>
+                      : rowChildren.reversed.toList() as List<Widget>,
             ),
-            wrap: theme.tapHeaderToExpand);
+            wrap: theme.tapHeaderToExpand!);
       }
     }
 
     Widget buildBody() {
-      Widget wrapBody(Widget child, bool tap) {
-        Alignment calcAlignment() {
+      Widget wrapBody(Widget? child, bool tap) {
+        Alignment? calcAlignment() {
           switch (theme.bodyAlignment) {
             case ExpandablePanelBodyAlignment.left:
               return Alignment.topLeft;
@@ -621,7 +625,7 @@ class ExpandablePanel extends StatelessWidget {
         }
 
         final widget = Align(
-          alignment: calcAlignment(),
+          alignment: calcAlignment()!,
           child: child,
         );
 
@@ -647,17 +651,17 @@ class ExpandablePanel extends StatelessWidget {
             );
           };
 
-      return builder(context, wrapBody(collapsed, theme.tapBodyToExpand),
-          wrapBody(expanded, theme.tapBodyToCollapse));
+      return builder(context, wrapBody(collapsed, theme.tapBodyToExpand!),
+          wrapBody(expanded, theme.tapBodyToCollapse!));
     }
 
     Widget buildWithHeader() {
       return Column(
         crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: <Widget>[
+        children: <Widget?>[
           buildHeaderRow(),
           buildBody(),
-        ],
+        ] as List<Widget>,
       );
     }
 
@@ -685,11 +689,11 @@ class ExpandablePanel extends StatelessWidget {
 /// An down/up arrow icon that toggles the state of [ExpandableController] when the user clicks on it.
 /// The model is accessed via [ScopedModelDescendant].
 class ExpandableIcon extends StatefulWidget {
-  final ExpandableThemeData _theme;
+  final ExpandableThemeData? _theme;
 
   ExpandableIcon({
-    @deprecated Color color,
-    ExpandableThemeData theme,
+    @deprecated Color? color,
+    ExpandableThemeData? theme,
     // ignore: deprecated_member_use_from_same_package
   }) : _theme = ExpandableThemeData.combine(
                 color != null ? ExpandableThemeData(iconColor: color) : null,
@@ -702,10 +706,10 @@ class ExpandableIcon extends StatefulWidget {
 
 class _ExpandableIconState extends State<ExpandableIcon>
     with SingleTickerProviderStateMixin {
-  AnimationController animationController;
-  Animation<double> animation;
-  ExpandableThemeData theme;
-  ExpandableController controller;
+  late AnimationController animationController;
+  late Animation<double> animation;
+  ExpandableThemeData? theme;
+  ExpandableController? controller;
 
   @override
   void initState() {
@@ -715,17 +719,17 @@ class _ExpandableIconState extends State<ExpandableIcon>
     animationController =
         AnimationController(duration: theme.animationDuration, vsync: this);
     animation = animationController.drive(Tween<double>(begin: 0.0, end: 1.0)
-        .chain(CurveTween(curve: theme.sizeCurve)));
+        .chain(CurveTween(curve: theme.sizeCurve!)));
     controller = ExpandableController.of(context, rebuildOnChange: false);
-    controller.addListener(_expandedStateChanged);
-    if (controller.expanded) {
+    controller!.addListener(_expandedStateChanged);
+    if (controller!.expanded) {
       animationController.value = 1.0;
     }
   }
 
   @override
   void dispose() {
-    controller.removeListener(_expandedStateChanged);
+    controller!.removeListener(_expandedStateChanged);
     animationController.dispose();
     super.dispose();
   }
@@ -739,11 +743,11 @@ class _ExpandableIconState extends State<ExpandableIcon>
   }
 
   _expandedStateChanged() {
-    if (controller.expanded &&
+    if (controller!.expanded &&
         const [AnimationStatus.dismissed, AnimationStatus.reverse]
             .contains(animationController.status)) {
       animationController.forward();
-    } else if (!controller.expanded &&
+    } else if (!controller!.expanded &&
         const [AnimationStatus.completed, AnimationStatus.forward]
             .contains(animationController.status)) {
       animationController.reverse();
@@ -758,8 +762,8 @@ class _ExpandableIconState extends State<ExpandableIcon>
     if (controller2 != controller) {
       controller?.removeListener(_expandedStateChanged);
       controller = controller2;
-      controller.addListener(_expandedStateChanged);
-      if (controller.expanded) {
+      controller!.addListener(_expandedStateChanged);
+      if (controller!.expanded) {
         animationController.value = 1.0;
       }
     }
@@ -772,21 +776,21 @@ class _ExpandableIconState extends State<ExpandableIcon>
     }
 
     return Padding(
-      padding: theme.iconPadding,
+      padding: theme!.iconPadding!,
       child: AnimatedBuilder(
         animation: animation,
         builder: (context, child) {
-          final showSecondIcon = theme.collapseIcon != theme.expandIcon &&
+          final showSecondIcon = theme!.collapseIcon != theme!.expandIcon &&
               animationController.value >= 0.5;
           return Transform.rotate(
-            angle: theme.iconRotationAngle *
+            angle: theme!.iconRotationAngle! *
                 (showSecondIcon
                     ? -(1.0 - animationController.value)
                     : animationController.value),
             child: Icon(
-              showSecondIcon ? theme.collapseIcon : theme.expandIcon,
-              color: theme.iconColor,
-              size: theme.iconSize,
+              showSecondIcon ? theme!.collapseIcon : theme!.expandIcon,
+              color: theme!.iconColor,
+              size: theme!.iconSize,
             ),
           );
         },
@@ -797,24 +801,24 @@ class _ExpandableIconState extends State<ExpandableIcon>
 
 /// Toggles the state of [ExpandableController] when the user clicks on it.
 class ExpandableButton extends StatelessWidget {
-  final Widget child;
+  final Widget? child;
 
-  ExpandableButton({@required this.child});
+  ExpandableButton({required this.child});
 
   @override
   Widget build(BuildContext context) {
     final controller = ExpandableController.of(context);
     final theme = ExpandableThemeData.withDefaults(null, context);
 
-    if (theme.useInkWell) {
+    if (theme.useInkWell!) {
       return InkWell(
-        onTap: controller.toggle,
+        onTap: controller!.toggle,
         child: child,
         borderRadius: theme.inkWellBorderRadius,
       );
     } else {
       return GestureDetector(
-        onTap: controller.toggle,
+        onTap: controller!.toggle,
         child: child,
       );
     }
@@ -836,15 +840,15 @@ class ScrollOnExpand extends StatefulWidget {
   /// If true then the widget will be scrolled to become visible when collapsed
   final bool scrollOnCollapse;
 
-  final ExpandableThemeData _theme;
+  final ExpandableThemeData? _theme;
 
   ScrollOnExpand({
-    Key key,
-    @required this.child,
+    Key? key,
+    required this.child,
     this.scrollOnExpand = true,
     this.scrollOnCollapse = true,
-    ExpandableThemeData theme,
-    @deprecated Duration scrollAnimationDuration,
+    ExpandableThemeData? theme,
+    @deprecated Duration? scrollAnimationDuration,
     // ignore: deprecated_member_use_from_same_package
   })  : _theme = scrollAnimationDuration != null
             ? ExpandableThemeData(
@@ -857,16 +861,16 @@ class ScrollOnExpand extends StatefulWidget {
 }
 
 class _ScrollOnExpandState extends State<ScrollOnExpand> {
-  ExpandableController _controller;
+  ExpandableController? _controller;
   int _isAnimating = 0;
-  BuildContext _lastContext;
-  ExpandableThemeData _theme;
+  BuildContext? _lastContext;
+  ExpandableThemeData? _theme;
 
   @override
   void initState() {
     super.initState();
     _controller = ExpandableController.of(context, rebuildOnChange: false);
-    _controller.addListener(_expandedStateChanged);
+    _controller!.addListener(_expandedStateChanged);
   }
 
   @override
@@ -875,9 +879,9 @@ class _ScrollOnExpandState extends State<ScrollOnExpand> {
     final newController =
         ExpandableController.of(context, rebuildOnChange: false);
     if (newController != _controller) {
-      _controller.removeListener(_expandedStateChanged);
+      _controller!.removeListener(_expandedStateChanged);
       _controller = newController;
-      _controller.addListener(_expandedStateChanged);
+      _controller!.addListener(_expandedStateChanged);
     }
   }
 
@@ -890,11 +894,11 @@ class _ScrollOnExpandState extends State<ScrollOnExpand> {
   _animationComplete() {
     _isAnimating--;
     if (_isAnimating == 0 && _lastContext != null && mounted) {
-      if ((_controller.expanded && widget.scrollOnExpand) ||
-          (!_controller.expanded && widget.scrollOnCollapse)) {
+      if ((_controller!.expanded && widget.scrollOnExpand) ||
+          (!_controller!.expanded && widget.scrollOnCollapse)) {
         _lastContext
             ?.findRenderObject()
-            ?.showOnScreen(duration: _theme.scrollAnimationDuration);
+            ?.showOnScreen(duration: _theme!.scrollAnimationDuration!);
       }
     }
   }
@@ -903,7 +907,7 @@ class _ScrollOnExpandState extends State<ScrollOnExpand> {
     if (_theme != null) {
       _isAnimating++;
       Future.delayed(
-          _theme.scrollAnimationDuration + Duration(milliseconds: 10),
+          _theme!.scrollAnimationDuration! + Duration(milliseconds: 10),
           _animationComplete);
     }
   }
